This module implements rasterization functionalites.
Rasterization is a widely used rendering technique to render
images of 3D scenes. which takes a scene described by geometric primitives as
input, and returns raster images of the visible part of the scene. Each point in
these images usually contains information about appearance or depth. A
rasterizer iterates over triangles by maintaining a depth buffer, which decides
the visibility of triangles.

PiperOrigin-RevId: 250963106