Make tf.graphics tests robust to when assertions are raised.
When an assert op is able to check the tensor's value statically
(e.g. the tensor is a constant), it marks this tensor as not feedable.
Then gradient checking ops try to feed a value and fail. The workaround
I used here is to wrap the constant tensors with an identity so that
asserts cannot evaluate them statically.

PiperOrigin-RevId: 262148668