Decreases random number of dimensions to a maximum of 3 to prevent huge tensors.
This fixes a bug where the test was not running with runs_per_test=2. Due to the random seed being fixed with runs_per_test=1 the numbers that were sampled were small enough by chance. With runs_per_test=2 a second set of tensor shapes was sampled which led to huge tensors that did not fit in memory.

PiperOrigin-RevId: 302609726