Fix missing tensorflow_graphics documentation.
Several modules were missing documentation due to missing imports in the __init__.py files. I added the ones I could see that were missing. I'm not sure if there is a systematic way to do this, as it seems we don't want to add every function to the API documentation. I omitted the rendering kernels and internal utility functions from the doc imports, for example.

PiperOrigin-RevId: 389260704