Switch from TFA resampler op to TFA interpolate_bilinear.
Maintaining the custom resampler op is a maintenance headache that requires aligning the versions of TFA and TFG for the opensource releases (e.g., https://github.com/tensorflow/graphics/issues/650). The interpolate_bilinear function is a pure-TF op that is likely somewhat slower, but has no maintenance problems. Using pure-TF ops also allows using higher-order derivatives and tf.ForwardAccumulator with the transform.sample function.

The implemented border behavior of interpolate_bilinear is DUPLICATE, whereas resampler is ZERO, so the sample function changes slightly to pad the image with zeros when BorderType.ZERO is requested.

PiperOrigin-RevId: 413971898